+++ database/migrations/2024_06_11_100603_create_purchase_order_items_table.php
@@ -21,7 +21,7 @@ return new class extends Migration
             $table->bigInteger('material_canvass_id');
             $table->bigInteger('material_id');
             $table->char('status',4)->default('PEND');
-            $table->decimal('quantity',10,2);
+            $table->double('quantity');
             $table->decimal('price',10,2);
         });
     }
